requirements.txt
openai==1.17.1
streamlit==1.35.0
langchain==0.1.20
langchain-community==0.0.38
langchain-core==0.1.52
langchain-experimental==0.0.58
langchain-openai==0.0.5
langchain-pinecone==0.1.0
langchain-text-splitters==0.0.2
python-dotenv==1.0.0